chore: remove unnecessary exporter file
